refactor route structure
