+++ backend/main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"backend/api"
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
 	"net/http"
@@ -30,7 +31,7 @@ func main() {
 		})
 	})
 
-	r.POST("/mortgage", calculateMortgage)
+	api.RegisterMortgageCalcRoute(r)
 
 	// Start server on port 8080
 	r.Run(":8080")
